backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-filter>=24.2
django-cors-headers>=4.3
drf-spectacular>=0.29.0,<0.30

# Auth
djangorestframework-simplejwt>=5.3

# Env & Config
# python-dotenv>=1.0
# django-environ>=0.11

# AI / NLP
# openai>=1.0
# langchain>=0.2
# tiktoken>=0.7

# Utils
Pillow>=10.0
requests>=2.31

# Dev tools
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
flake8>=7.0
black>=24.0
isort>=5.13

# # Production
# gunicorn>=22.0
# whitenoise>=6.6